Backend/#40/internal js docs+readme (#43)
* did jsdocs for dbapi

* added a README file

* added fullstops for formattig

* removed db.ts since the testing it did is no longer needed

* edited the file structure and started adding more to the README.md

* README formatting

* Edited readme file
