+++ server/controllers/api.ts
@@ -1,7 +1,7 @@
 import { Response, Request } from 'express'
 import * as util from 'util'
 
-import Database from '../dbApi'
+import Database from '../src/database'
 
 interface IGetCourseParams {
   termId: string
@@ -10,9 +10,9 @@ interface IGetCourseParams {
 
 /**
  * GET /api/terms/:termId/courses/:courseId/
+ * termId expected in yyyy-term format
  */
 export const getCourse = async (req: Request, res: Response) => {
-  //termId expected in yyyy-term format
   const params: IGetCourseParams = req.params
   const [year, term] = params.termId.split('-')
   const course = await Database.dbRead(year, term, params.courseId)
@@ -26,6 +26,10 @@ export const getCourse = async (req: Request, res: Response) => {
   }
 }
 
+/**
+ * GET /api/terms/:termId/courses
+ * termId expected in yyyy-term format
+ */
 export const getCourseList = async (req: Request, res: Response) => {
   const params: IGetCourseParams = req.params
   const [year, term] = params.termId.split('-')
